requirements.txt
aiogram
python-dotenv
watchfiles
aiohttp
pdf417gen
Pillow
requests
